shift get-token method to auth-service file
